+++ src/components/post-list-item/index.scss
@@ -1,0 +1,46 @@
+@import "../../style/valuables.scss";
+
+.blog-post-preview {
+  background: white;
+  box-shadow: $shadow;
+
+  padding: 0.2rem 1.6rem;
+  margin: 0.6rem 0rem;
+
+  & .title {
+    font-size: 1.3rem;
+    margin: 0rem;
+  }
+
+  & .summary {
+    font-size: 0.8rem;
+    color: #7d7d7d;
+    margin-bottom: 0.7rem;
+  }
+
+  & .tag-container {
+    margin-bottom: 0.7rem;
+
+    & .tag {
+      margin: 0rem 0.2rem;
+      padding: 0.3rem;
+      border: 1.2px solid #555;
+      border-radius: 10px;
+      font-size: 12px;
+
+      &:hover {
+        color: white;
+        font-weight: 700;
+        background-color: #77af9c;
+        border: none;
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 640px) {
+  .blog-post-preview {
+    margin: 0.1rem 0rem;
+    box-shadow: none;
+  }
+}
