+++ src/style/_valuables.scss
@@ -1,7 +1,7 @@
 $header-height: 3.5rem;
 $footer-height: 3.5rem;
 
-$shadow: 0px 2px 5px 2px rgba(0, 0, 0, 0.6);
+$shadow: 0rem 0.2rem 0.2rem 0.01rem rgba(0, 0, 0, 0.35);
 $shadowHover: 0px 2px 5px 2px rgba(0, 0, 0, 0.6);
 
 $avataImg: 130px;
