+++ src/style/posts.scss
@@ -5,80 +5,7 @@
   margin: 1rem auto;
   margin-top: 1rem;
 
-  max-width: 900px;
-  font-family: "NanumSquareRound";
-
-  & .blog-post-preview {
-    display: flex;
-    flex-direction: column;
-
-    background: white;
-    box-shadow: 0px 5px 10px 1px #aaaaaa;
-
-    padding: 1.5rem 2rem;
-    margin: 0.7rem 0rem;
-
-    & :last-child {
-      border-bottom-width: 0;
-    }
-
-    & .blog-container {
-      display: inline-block;
-      //flex-direction: row;
-
-      & .title {
-        width: 100%;
-        margin-bottom: 1.5rem;
-
-        font: {
-          family: "SCDream";
-          size: 24px;
-          weight: 700;
-        }
-      }
-
-      & .blog-sub-container {
-        display: inline-block;
-
-        max-width: 80%;
-
-        & .summary {
-          padding-right: 1rem;
-          margin-bottom: 1rem;
-        }
-
-        & .author,
-        .date {
-          font: {
-            size: 0.9rem;
-          }
-
-          text-align: center;
-          margin-bottom: 0px;
-        }
-
-        & .tagContainer {
-          & .tag {
-            border: 1px solid #555;
-            border-radius: 3px;
-
-            font: {
-              size: 12px;
-            }
-
-            margin: 0rem 0.5rem;
-            padding: 0.1rem 0.3rem;
-          }
-          & :first-child {
-            margin-left: 0rem;
-          }
-        }
-        & .link {
-          margin: 0 auto;
-        }
-      }
-    }
-  }
+  max-width: 700px;
 }
 
 .right {
@@ -103,66 +30,6 @@
 @media only screen and (max-width: 640px) {
   .blog-posts {
     margin: 0px;
-    // margin-top: $header-height + 1rem;
-
-    & .blog-post-preview {
-      padding: 1rem;
-      margin: 0.3rem 0rem;
-      box-shadow: none;
-
-      & .blog-container {
-        display: flex;
-        flex-direction: row;
-
-        & .title {
-          font: {
-            size: 16px;
-          }
-          margin-bottom: 0.5rem;
-        }
-
-        & .blog-sub-container {
-          max-width: 100%;
-
-          & .summary {
-            font: {
-              size: 13px;
-            }
-            margin-bottom: 0.3rem;
-          }
-
-          & .date {
-            font: {
-              size: 12px;
-            }
-          }
-
-          & .tagContainer {
-            & .tag {
-              border: 1px solid #555;
-              border-radius: 3px;
-
-              font: {
-                size: 13px;
-              }
-
-              margin: 0rem 0.5rem;
-              padding: 0.1rem 0.3rem;
-            }
-            & :first-child {
-              margin-left: 0rem;
-            }
-          }
-          & .link {
-            margin: 0 auto;
-          }
-        }
-
-        & .blog-sub-container:last-child {
-          display: none;
-        }
-      }
-    }
   }
 
   .posts-bottom {
