+++ src/style/tag.scss
@@ -1,19 +1,19 @@
-@import "font.scss";
-@import "valuables.scss";
+// @import "font.scss";
+// @import "valuables.scss";
 
-.tagContainer{
-  & .tag{
-    border: 1px solid #555;
-    border-radius: 3px;
+// .tagContainer {
+//   & .tag {
+//     border: 1px solid #555;
+//     border-radius: 3px;
 
-    font: {
-      size: 15px;
-    }
+//     font: {
+//       size: 15px;
+//     }
 
-    margin: 0rem 0.5rem;
-    padding: 0.1rem 0.3rem;
-  }
-  & :first-child{
-    margin-left: 0rem;
-  }
-}
+//     margin: 0rem 0.5rem;
+//     padding: 0.1rem 0.3rem;
+//   }
+//   & :first-child {
+//     margin-left: 0rem;
+//   }
+// }
